Implement deployment id, to avoid caching issues
I use the unix epoch as a key to prepend to filenames that should not be
cached in the next deploy. It's not optimal, but it's better than
nothing in this early phase of development.

Also, it really needs refactoring. There's a lot of copy and paste that
I haven't cleaned up yet.