Stop using Travis-CI for now
It is kind of broken, since we're using time 1.5 now, which requires ghc
7.10.*, and it is not officially supported by Travis. Besides, hs-tls is
also kind of broken, which is why it can't check URL's like
https://seismesesnoporto.ga, or https://sqitch.org. It is marking those
as unavailable, simply because it can't properly connect. So `cabal run
check` is failing, even though everything is on the right place.

When hs-tls is fixed, and when Travis support ghc 7.10, I'll come back
:)